fix adding new message templates
