+++ src/components/App/App.css
@@ -1,3 +1,10 @@
-.form-container {
-  width: 500px;
+select {
+  width: 200px;
+}
+
+.message-container {
+  height: 100px;
+  border: 1px solid black;
+  padding: 10px;
+  width: 700px;
 }
